@@ -4,7 +4,7 @@
 
 	 <groupId>io.omengye</groupId>
 	 <artifactId>wscloud</artifactId>
-	 <version>0.0.1-SNAPSHOT</version>
+	 <version>0.0.2</version>
 	 <packaging>pom</packaging>
 
 	 <name>wscloud</name>
@@ -13,9 +13,9 @@
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-		<java.version>1.8</java.version>
-		<spring-cloud.version>Hoxton.SR6</spring-cloud.version>
-		<spring-boot.version>2.2.8.RELEASE</spring-boot.version>
+		<java.version>11</java.version>
+		<spring-cloud.version>2021.0.4</spring-cloud.version>
+		<spring-boot.version>2.7.4</spring-boot.version>
 		<project.version>0.0.1-SNAPSHOT</project.version>
 		<docker-maven-plugin.version>0.33.0</docker-maven-plugin.version>
 		<dockerHost.url>tcp://127.0.0.1:2375</dockerHost.url>
@@ -50,12 +50,6 @@
 				<groupId>org.projectlombok</groupId>
 				<artifactId>lombok</artifactId>
 				<version>1.18.10</version>
-			</dependency>
-
-			<dependency>
-				<groupId>com.nimbusds</groupId>
-				<artifactId>nimbus-jose-jwt</artifactId>
-				<version>8.10</version>
 			</dependency>
 
 			<dependency>
