@@ -1,9 +1,20 @@
 package io.omengye.gcs.configure;
 
+import com.nimbusds.jose.jwk.JWKSet;
+import com.nimbusds.jose.jwk.source.ImmutableJWKSet;
+import com.nimbusds.jose.proc.JWKSecurityContext;
+import com.nimbusds.jose.proc.JWSVerificationKeySelector;
+import com.nimbusds.jwt.proc.ConfigurableJWTProcessor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.config.annotation.web.reactive.EnableWebFluxSecurity;
 import org.springframework.security.config.web.server.ServerHttpSecurity;
+import org.springframework.security.oauth2.jose.jws.SignatureAlgorithm;
+import org.springframework.security.oauth2.jwt.NimbusReactiveJwtDecoder;
 import org.springframework.security.web.server.SecurityWebFilterChain;
+import reactor.core.publisher.Flux;
+
+import java.util.function.Consumer;
 
 @EnableWebFluxSecurity
 public class SecurityConfig {
@@ -13,11 +24,11 @@ public class SecurityConfig {
 		http
 			.authorizeExchange()
 				.pathMatchers("/actuator/**").permitAll()
-				.pathMatchers("/api/**").hasAuthority("SCOPE_message:read")
 				.anyExchange().authenticated()
 				.and()
-			.oauth2ResourceServer()
+				.oauth2ResourceServer()
 				.jwt();
 		return http.build();
 	}
+
 }
