@@ -3,12 +3,13 @@ package io.omengye.gcs.entity;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import io.omengye.common.utils.Utils;
 import io.omengye.gcs.valid.CollectionValid;
+import jakarta.validation.constraints.Max;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
 import lombok.Data;
 
-import javax.validation.constraints.Max;
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
+
 
 @Data
 @JsonIgnoreProperties(ignoreUnknown=true)
