@@ -1,5 +1,6 @@
 package io.omengye.gcs.exception;
 
+import org.springframework.boot.web.error.ErrorAttributeOptions;
 import org.springframework.boot.web.reactive.error.DefaultErrorAttributes;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
@@ -11,10 +12,8 @@ import java.util.Map;
 public class GlobalErrorAttributes extends DefaultErrorAttributes {
 
     @Override
-    public Map<String, Object> getErrorAttributes(ServerRequest request,
-                                                  boolean includeStackTrace) {
-        Map<String, Object> map = super.getErrorAttributes(
-                request, includeStackTrace);
+    public Map<String, Object> getErrorAttributes(ServerRequest request, ErrorAttributeOptions options) {
+        Map<String, Object> map = super.getErrorAttributes(request, options);
         map.put("status", HttpStatus.BAD_REQUEST);
         return map;
     }
