@@ -1,15 +1,16 @@
 package io.omengye.gcs.service;
 
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
+
+import javax.annotation.Resource;
 
 @Component
 public class StratRunner implements CommandLineRunner {
 
-    @Autowired
-    ChooseItemService chooseItemService;
+    @Resource
+    private ChooseItemService chooseItemService;
 
     @Override
     public void run(String... args) throws Exception {
