@@ -1,9 +1,10 @@
 package io.omengye.gcs.valid;
 
 import io.omengye.common.utils.Utils;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
 
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
+
 import java.util.Arrays;
 import java.util.List;
 
