@@ -6,8 +6,9 @@ public class Constants {
 
 	public static final int EXPIRE_TOKEN_TIME = 9*3600;
 	public static final int EXPIRE_DELAY_TOKEN_TIME = 5*60;
-	public static final int EXPIRE_REFRESH_TOKEN_TIME = 12*3600;
 
 	public static final int MAX_VISIT = 2000;
+
+	public static final String DEFAULT_CLIENT_CONFIGURATION_AUTHORIZED_SCOPE = "message.read";
 
 }
